% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_nuevas_provisionales.R
\name{DemBas_indicadores_demograficos}
\alias{DemBas_indicadores_demograficos}
\title{Función "todo en uno": calcula indicadores demográficos completos}
\usage{
DemBas_indicadores_demograficos(datos, nacimientos = NA, l0 = 1e+05)
}
\arguments{
\item{datos}{Data frame con columnas: edad, sexo, poblacion, defunciones (opcional)}

\item{nacimientos}{Número anual de nacimientos}

\item{l0}{Raíz de la tabla de vida (defecto 100000)}
}
\value{
Data frame resumen
}
\description{
Función "todo en uno": calcula indicadores demográficos completos
}
\details{
Retorna un resumen con:
- Tasas brutas (natalidad, mortalidad)
- Crecimiento natural
- Indicadores estructurales
- Esperanza de vida (si hay defunciones)
}
