% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_nuevas_provisionales.R
\name{DemBas_indicadores_estructura}
\alias{DemBas_indicadores_estructura}
\title{Calcula indicadores estructurales de una población}
\usage{
DemBas_indicadores_estructura(datos, grupos_edad = c(0, 15, 65, Inf))
}
\arguments{
\item{datos}{Data frame con columnas: edad, sexo, poblacion}

\item{grupos_edad}{Vector con límites de grupos de edad (ej: c(0,15,65,Inf))}
}
\value{
Data frame con indicadores
}
\description{
Calcula indicadores estructurales de una población
}
\details{
Calcula:
- Ratio de dependencia (jóvenes y ancianos)
- Índice de envejecimiento
- Razón de masculinidad
- % población por grandes grupos
}
