% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piramidespoblacion.R
\name{DemBas_piramidePorc_Generaciones}
\alias{DemBas_piramidePorc_Generaciones}
\title{DemBas_piramidePorc_Generaciones}
\usage{
DemBas_piramidePorc_Generaciones(
  pop3,
  Ano_ref = 2020,
  Gtitulo = "Pirámide Población de la provincia de Sevilla",
  Gsubtitulo = "Año 2020",
  Gtitulo.X = "Porcentajes",
  GHombresEtiq = "Hombres",
  GMujeresEtiq = "Mujeres",
  Gedadfinal = 100,
  Gext_izq = -4.5,
  Gext_der = 4.5,
  Glimite = 0.5,
  Gsizeletra = 2.5,
  GpresentaResumen = TRUE,
  GSegmentos = TRUE,
  GHombresColor = "#5BB4E5",
  GMujeresColor = "#DE61D8",
  GSegmentosColor = "#00ff00",
  Gguardar = FALSE,
  Garchivo = "piramide.png",
  Gwidth = 12,
  Gheight = 10
)
}
\arguments{
\item{pop3}{data.frame o tibble with columns: Edad, Sexo, Porcentajes}

\item{Gheight}{}
}
\description{
DemBas_piramidePorc_Generaciones
}
\examples{
load(file = system.file("examples/04003px.RData", package = "DemographyBasic"))
datosPiramide =  datos |>
  dplyr::filter(Ano == 2020 &
                  Sexo \%in\% c("Mujeres", "Hombres") &
                  Edad != "TOTAL" &
                  CCAA.Prov == "Sevilla" &
                  Espanoles.Extranjeros == "Españoles") |>
  dplyr::rename(Poblacion = value) |>
  dplyr::select(Edad, Sexo, Poblacion)

g_pir1 = DemBas_piramidePorc_Generaciones(datosPiramide,
                                          Gtitulo = "Pirámide Población de la provincia de Sevilla",
                                          Gsubtitulo = "Año 2020  (españoles)")
g_pir1
g_pir2 = DemBas_piramidePorc_Generaciones(datosPiramide,
                                          Gtitulo = "Pirámide Población de la provincia de Sevilla",
                                          Gsubtitulo = "Año 2020  (españoles)",
                                          GSegmentos = FALSE, GpresentaResumen = FALSE)
g_pir2
}
