% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablavida.R
\name{DemBas_tablavida_abreviada2}
\alias{DemBas_tablavida_abreviada2}
\title{Calcula la tabla de vida abreviada para edades agrupadas}
\usage{
DemBas_tablavida_abreviada2(mx)
}
\arguments{
\item{mx}{Vector numérico con las tasas centralizadas de mortalidad para
cada grupo de edad. Los grupos deben corresponderse con: 0, 1-4, 5-9,
10-14, ..., 95-99, 100+.}
}
\value{
Data frame con la tabla de vida abreviada que contiene las columnas:
  \describe{
    \item{x}{Edad inicial de cada grupo (0, 1, 5, 10, ..., 100)}
    \item{mx}{Tasa centralizada de mortalidad (redondeada a 5 decimales)}
    \item{qx}{Probabilidad de muerte en el intervalo n}
    \item{px}{Probabilidad de sobrevivir en el intervalo n}
    \item{lx}{Supervivientes al inicio del intervalo}
    \item{dx}{Defunciones en el intervalo}
    \item{Lx}{Personas-año vividas en el intervalo}
    \item{Tx}{Personas-año restantes desde el inicio del intervalo}
    \item{ex}{Esperanza de vida a la edad x}
    \item{Sx}{Fracción de supervivencia del intervalo}
  }
}
\description{
Construye una tabla de vida abreviada a partir de las tasas de mortalidad
para edades agrupadas (0, 1-4, 5-9, 10-14, ...). El método aplica
coeficientes de separación de Greville para el cálculo de las personas-año
vividas (nLx).
}
\examples{
# Tasas de mortalidad para España 2003
mx0 <- 1733 / 441881
mx <- c(mx0, 0.00027, 0.00013, 0.00016, 0.00043, 0.00057, 0.00059,
        0.00081, 0.00115, 0.00174, 0.00258, 0.00376, 0.00569, 0.00818,
        0.01346, 0.02206, 0.03844, 0.06981, 0.12872, 0.21674, 0.31705,
        0.48258)
tv <- DemBas_tablavida_abreviada2(mx)
tv

}
\references{
Greville, T. N. E. (1948). Mortality tables analyzed by cause of death.
  The American Statistician, 2(4), 23-27.
}
