% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablavida.R
\name{DemBas_tablavida_abreviada_calculadora}
\alias{DemBas_tablavida_abreviada_calculadora}
\title{Calcula la tabla de vida abreviada con redondeo manual}
\usage{
DemBas_tablavida_abreviada_calculadora(nMx, l0 = 1e+05, muestraSx = TRUE)
}
\arguments{
\item{nMx}{Vector numérico con las tasas centralizadas de mortalidad para
cada grupo de edad quinquenal.}

\item{l0}{Valor numérico con la población inicial (radix). Por defecto 100000.}

\item{muestraSx}{Valor lógico. Si TRUE, incluye la columna Sx en el resultado.}
}
\value{
tibble con la tabla de vida abreviada calculada con redondeo manual.
}
\description{
Versión de \code{\link{DemBas_tablavida_abreviada}} que simula el proceso
de redondeo manual típico de los cálculos demográficos tradicionales.
Redondea cada columna de forma independiente en cada paso, imitando
el proceso que se realizaría con una calculadora.
}
\examples{
mx0 <- 1733 / 441881
mx <- c(mx0, 0.00027, 0.00013, 0.00016, 0.00043, 0.00057, 0.00059,
        0.00081, 0.00115, 0.00174, 0.00258, 0.00376, 0.00569, 0.00818,
        0.01346, 0.02206, 0.03844, 0.06981, 0.12872, 0.21674, 0.31705,
        0.48258)
tv <- DemBas_tablavida_abreviada_calculadora(mx)
tv

}
\seealso{
\code{\link{DemBas_tablavida_abreviada}} para la versión sin
  redondeo manual.
}
