% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablavida.R
\name{f_tb_qx}
\alias{f_tb_qx}
\alias{f_tb_px}
\alias{f_tb_lx}
\alias{f_tb_dx}
\alias{f_tb_Lx}
\alias{f_tb_Tx}
\alias{f_tb_ex}
\alias{f_tb_Sx}
\title{Funciones auxiliares para cálculo de tablas de vida}
\usage{
f_tb_qx(Mx)

f_tb_px(qx)

f_tb_lx(px, l0 = 1e+05)

f_tb_dx(lx, qx)

f_tb_Lx(lx, dx, Mx)

f_tb_Tx(Lx)

f_tb_ex(Tx, lx)

f_tb_Sx(Lx, lx)
}
\arguments{
\item{Mx}{Vector de tasas de mortalidad (para f_tb_Lx).}

\item{qx}{Vector de probabilidades de muerte (para f_tb_dx).}

\item{px}{Vector de probabilidades de supervivencia.}

\item{l0}{Población inicial o radix (para f_tb_lx). Por defecto 100000.}

\item{lx}{Vector de supervivientes (para f_tb_ex, f_tb_Sx).}

\item{dx}{Vector de defunciones en el intervalo.}

\item{Lx}{Vector de personas-año vividas (para f_tb_Tx, f_tb_ex, f_tb_Sx).}
}
\description{
Estas funciones calculan diferentes columnas de una tabla de vida
a partir de las tasas de mortalidad centralizadas.
}
