% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablavida.R
\name{f_tba_nqx}
\alias{f_tba_nqx}
\alias{f_tba_npx}
\alias{f_tba_lx}
\alias{f_tba_ndx}
\alias{f_tba_nLx}
\alias{f_tba_Tx}
\alias{f_tba_ex}
\alias{f_tba_Sx}
\title{Funciones auxiliares para cálculo de tablas de vida abreviadas}
\usage{
f_tba_nqx(nMx, vn)

f_tba_npx(nqx)

f_tba_lx(npx, l0 = 1e+05)

f_tba_ndx(lx, nqx)

f_tba_nLx(lx, nMx, vn)

f_tba_Tx(nLx)

f_tba_ex(Tx, lx)

f_tba_Sx(Lx, lx)
}
\arguments{
\item{nMx}{Vector de tasas de mortalidad (para f_tba_nLx).}

\item{vn}{Vector con las longitudes de cada intervalo de edad.}

\item{nqx}{Vector de probabilidades de muerte (para f_tba_ndx).}

\item{npx}{Vector de probabilidades de supervivencia (para f_tba_lx).}

\item{l0}{Población inicial o radix (para f_tba_lx). Por defecto 100000.}

\item{lx}{Vector de supervivientes (para f_tba_ex, f_tba_Sx).}

\item{nLx}{Vector de personas-año vividas (para f_tba_Tx, f_tba_ex, f_tba_Sx).}

\item{Lx}{Vector de personas-año vividas (para f_tba_Sx).}
}
\description{
Estas funciones calculan diferentes columnas de una tabla de vida
abreviada a partir de las tasas de mortalidad centralizadas para
grupos de edad.
}
