% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demografia.R
\name{DemBas_agrupar_variable}
\alias{DemBas_agrupar_variable}
\title{DemBas_agrupar_variable}
\usage{
DemBas_agrupar_variable(
  variable,
  metodo = 1,
  final = 85,
  vbreaks = NULL,
  vlabels = NULL,
  labelfinal = NULL
)
}
\arguments{
\item{labelfinal}{}
}
\description{
DemBas_agrupar_variable
}
