% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demografia.R
\name{DemBas_anade_GEdad5}
\alias{DemBas_anade_GEdad5}
\title{DemBas_anade_GEdad5}
\usage{
DemBas_anade_GEdad5(datos, varEdad)
}
\arguments{
\item{datos}{data.frame o tibble con una variable de edad}

\item{varEdad}{debe ser de tipo: numeric o integer}
}
\description{
Añade una variable de grupos quinquenales de edad
}
\examples{
datos = DemBas_read_px(system.file("examples/9663.px", package = "DemographyBasic"))
datosPob = datos |>
  dplyr::filter(Periodo == "1 de enero de  2018",
                Edad.simple != "Total",
                Sexo != "Ambos sexos") |>
  dplyr::select(Sexo, Edad.simple, Poblacion = value)

datosPob2 = datosPob |>
  dplyr::mutate(
    Edad = as.numeric(gsub("[años|año]", "", Edad.simple)),
    Poblacion = round(Poblacion, 0)
  )

datosPob2_conGruposEdad = DemBas_anade_GEdad5(datosPob2, Edad)
head(datosPob2_conGruposEdad,15)


datosPirAgru = DemBas_anade_GEdad5(datosPob2, Edad) |>
  dplyr::select(Sexo, Edad, Poblacion, GEdad5) |>
  dplyr::group_by(GEdad5, Sexo) |>
  dplyr::summarise(Poblacion = sum(Poblacion), .groups = "keep")

head(datosPirAgru)

}
