% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demografia.R
\name{DemBas_asigna_codigos_CCAA}
\alias{DemBas_asigna_codigos_CCAA}
\title{DemBas_asigna_codigos_CCAA}
\usage{
DemBas_asigna_codigos_CCAA(columna, columna_codigos, etiqueta, nuevocodigo)
}
\arguments{
\item{columna}{}

\item{columna_codigos}{}

\item{etiqueta}{}

\item{nuevocodigo}{}
}
\description{
DemBas_asigna_codigos_CCAA
}
