% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demografia.R
\name{DemBas_crea_colores_brewer}
\alias{DemBas_crea_colores_brewer}
\title{DemBas_crea_colores_brewer}
\usage{
DemBas_crea_colores_brewer(cuantos, que_paletacolor = 3)
}
\arguments{
\item{que_paletacolor}{}
}
\description{
crea vector de colores con paleta brewer
}
