% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piramidespoblacion.R
\name{DemBas_datos_piramidePorc}
\alias{DemBas_datos_piramidePorc}
\title{DemBas_datos_piramidePorc}
\usage{
DemBas_datos_piramidePorc(
  datosPiramide,
  GEdad_final = 100,
  etiq.hombre = "Hombres"
)
}
\arguments{
\item{datosPiramide}{}

\item{GEdad_final}{}

\item{etiq.hombre}{}
}
\description{
DemBas_datos_piramidePorc
}
