% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demografia.R
\name{DemBas_dem_EdadMedia_t}
\alias{DemBas_dem_EdadMedia_t}
\title{DemBas_dem_EdadMedia_t}
\usage{
DemBas_dem_EdadMedia_t(vTasasEspecificas_x_t, Edad.marcas)
}
\arguments{
\item{Edad.marcas}{}
}
\description{
DemBas_dem_EdadMedia_t
}
