% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demografia.R
\name{DemBas_dem_ISinteticoCoyuntural_t}
\alias{DemBas_dem_ISinteticoCoyuntural_t}
\title{DemBas_dem_TasaEspecificaEdad_t}
\usage{
DemBas_dem_ISinteticoCoyuntural_t(vTasasEspecificas_x_t)
}
\arguments{
\item{vTasasEspecificas_x_t}{}
}
\description{
DemBas_dem_TasaEspecificaEdad_t
}
