% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demografia.R
\name{DemBas_dem_TasaBruta_t}
\alias{DemBas_dem_TasaBruta_t}
\title{DemBas_dem_TasaBruta_t}
\usage{
DemBas_dem_TasaBruta_t(NumEventos_t, PobMedia_t)
}
\arguments{
\item{PobMedia_t}{}
}
\description{
DemBas_dem_TasaBruta_t
}
