% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demografia.R
\name{DemBas_dem_TasaEspecifica_t}
\alias{DemBas_dem_TasaEspecifica_t}
\title{DemBas_dem_TasaEspecifica_t}
\usage{
DemBas_dem_TasaEspecifica_t(vNumEventos_x_t, vPobMedia_x_t)
}
\arguments{
\item{vPobMedia_x_t}{}
}
\description{
DemBas_dem_TasaEspecifica_t
}
