% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demografia.R
\name{DemBas_dem_TasaGeneral_t}
\alias{DemBas_dem_TasaGeneral_t}
\title{DemBas_dem_TasaEspecifica_t}
\usage{
DemBas_dem_TasaGeneral_t(NumEventos_t, PobMedia_Referida_t)
}
\arguments{
\item{NumEventos_t}{}

\item{PobMedia_Referida_t}{}
}
\description{
DemBas_dem_TasaEspecifica_t
}
