% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demografia.R
\name{DemBas_etiquetas_gruposEdad}
\alias{DemBas_etiquetas_gruposEdad}
\title{DemBas_etiquetas_gruposEdad}
\usage{
DemBas_etiquetas_gruposEdad(metodo = 1, final = 85, labelfinal = NULL)
}
\arguments{
\item{metodo}{}

\item{final}{}

\item{labelfinal}{}
}
\description{
DemBas_etiquetas_gruposEdad
}
