% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demografia.R
\name{DemBas_extrae_codigo_CCAA}
\alias{DemBas_extrae_codigo_CCAA}
\title{DemBas_extrae_codigo_CCAA}
\usage{
DemBas_extrae_codigo_CCAA(columna)
}
\arguments{
\item{columna}{}
}
\description{
DemBas_extrae_codigo_CCAA
}
