% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demografia.R
\name{DemBas_extrae_codigo_ccaa}
\alias{DemBas_extrae_codigo_ccaa}
\title{DemBas_extrae_codigo_ccaa}
\usage{
DemBas_extrae_codigo_ccaa(vCCAA, ConvierteCodSIANE = TRUE)
}
\arguments{
\item{vCCAA}{}

\item{ConvierteCodSIANE}{}
}
\description{
DemBas_extrae_codigo_ccaa
}
