% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demografia.R
\name{DemBas_extrae_codigo_provincia}
\alias{DemBas_extrae_codigo_provincia}
\title{DemBas_extrae_codigo_provincia}
\usage{
DemBas_extrae_codigo_provincia(vprovincias)
}
\arguments{
\item{vprovincias}{}
}
\value{
vector character
}
\description{
DemBas_extrae_codigo_provincia
}
