% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demografia.R
\name{DemBas_extrae_notienen_codigos}
\alias{DemBas_extrae_notienen_codigos}
\title{DemBas_extrae_notienen_codigos}
\usage{
DemBas_extrae_notienen_codigos(columna, columnacodigos)
}
\arguments{
\item{columna}{}

\item{columnacodigos}{}
}
\description{
DemBas_extrae_notienen_codigos
}
