% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxR2.R
\name{DemBas_import_px_IECA}
\alias{DemBas_import_px_IECA}
\title{DemBas_import_px_IECA}
\usage{
DemBas_import_px_IECA(id, ficheropx = NULL)
}
\arguments{
\item{id}{character with the id of the px file in the IECA web}

\item{ficheropx}{character with the name of the file px to save.
If is NULL, a temporal file is created}
}
\value{
data.frame with the data
}
\description{
Importa un fichero px de la web del IECA a partir de su identificador
}
\examples{
df103965 = DemBas_import_px_IECA("103965")

df103965 = DemBas_import_px_IECA("103965", "ieca_export_103965_ff4.px")

}
