% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxR2.R
\name{DemBas_import_px_INE}
\alias{DemBas_import_px_INE}
\title{DemBas_import_px_INE}
\usage{
DemBas_import_px_INE(id, ficheropx = NULL)
}
\arguments{
\item{id}{character with the id of the px file in the INE web}

\item{ficheropx}{character with the name of the file px to save.
If is NULL, a temporal file is created}
}
\value{
data.frame with the data
}
\description{
Importa un fichero px de la web del INE a partir de su identificador
}
\examples{
df2855_ine = DemBas_import_px_INE("2855")

df2855_ine2 = DemBas_import_px_INE("2855", "2855_ine.px")
}
