% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxR2.R
\name{DemBas_leer_metadatos_px_IECA}
\alias{DemBas_leer_metadatos_px_IECA}
\title{DemBas_leer_metadatos_px_IECA}
\usage{
DemBas_leer_metadatos_px_IECA(ficheropx)
}
\arguments{
\item{ficheropx}{character with the name of the file px}
}
\value{
data.frame with the metadata, column 1 is the name of the variable and column 2 is the content
}
\description{
Leer metadatos de un fichero px del IECA
}
\examples{
df103965 = DemBas_import_px_IECA("103965", "ieca_export_103965_ff4.px")
df103965_meta = DemBas_leer_metadatos_px_IECA("ieca_export_103965_ff4.px")
}
