% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxR2.R
\name{DemBas_leer_metadatos_px_INE}
\alias{DemBas_leer_metadatos_px_INE}
\title{DemBas_leer_metadatos_px_INE}
\usage{
DemBas_leer_metadatos_px_INE(ficheropx)
}
\arguments{
\item{ficheropx}{character with the name of the file px}
}
\value{
data.frame with the metadata, column 1 is the name of the variable and column 2 is the content
}
\description{
Leer metadatos de un fichero px del INE
}
\examples{
df2855 = DemBas_import_px_INE("2855", "2855_ine.px")
df2855_meta = DemBas_leer_metadatos_px_INE("ieca_export_103965_ff4.px")
}
