% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablavida.R
\name{DemBas_mx}
\alias{DemBas_mx}
\title{Calcula las tasas especificas de mortalidad para edades simples}
\usage{
DemBas_mx(Px, Dx, N0, D0)
}
\arguments{
\item{Px}{vector de población media por edad}

\item{Dx}{vector de defunciones por edad}

\item{N0}{nacimientos en el periodo}

\item{D0}{defunciones en el periodo}
}
\value{
devuelve un vector con las tasas especificas de mortalidad
}
\description{
Calcula las tasas especificas de mortalidad para edades simplex a partir de las defunciones y población
}
