% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablavida.R
\name{DemBas_mx_abreviada}
\alias{DemBas_mx_abreviada}
\title{Calcula las tasas especificas de mortalidad para edades agrupadas}
\usage{
DemBas_mx_abreviada(Px, Dx, N0, D0)
}
\arguments{
\item{Px}{vector de poblaci칩n media por edad}

\item{Dx}{vector de defunciones por edad}

\item{N0}{nacimientos en el periodo}

\item{D0}{defunciones en el periodo}
}
\value{
devuelve un vector con las tasas especificas de mortalidad
}
\description{
Calcula las tasas especificas de mortalidad para edades agrupadas a partir de las defunciones y poblaci칩n
}
\examples{
x = c(0,1,seq(5,100,by=5))
Px = c(NA,1627456,1938350,2104636,2388049,3070467,3614444,3545550,
       3431304,3182840,2791972,2498361,2334676,1953022,1978465,
       1898370,1492487,974162,495260,203924,46078,5139)
Dx = c(NA,442,254,341,1032,1746,2136,2872,
       2933,5545,7193,9401,13294,15972,26636,
       41879,57377,68007,63007,44198,14609,2480)
N0 = 441881 # Nacimientos en Espa침a en 2003: 441881
D0 = 1733 # Defunciones de menores de un a침o durante 2003: 1733
mx = DemBas_mx_abreviada(Px,Dx,N0,D0)
mx
}
