% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxR2.R
\name{DemBas_num_decimales_px_IECA}
\alias{DemBas_num_decimales_px_IECA}
\title{DemBas_num_decimales_px_IECA}
\usage{
DemBas_num_decimales_px_IECA(v1, num_decimales = 2)
}
\arguments{
\item{v1}{vector numérico a redondear}

\item{num_decimales}{número de decimales especificados en el fichero px de metadatos del IECA}
}
\description{
Redondea un vector numérico para corregir el número de decimales que viene de un fichero px del IECA
}
\examples{
df_19824 = DemBas_import_px_IECA("19824", "ieca_19824.px")
df_19824_meta = DemBas_leer_metadatos_px_IECA("ieca_19824.px")
df_19824_meta[df_19824_meta$Claves \%in\% c("DECIMALS", "SHOWDECIMALS"),]
df_19824$value = DemBas_num_decimales_px_IECA(df_19824$value)
head(df_19824,15)
}
