% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piramidespoblacion.R
\name{DemBas_piramide_compuestasCateg_ggplot2}
\alias{DemBas_piramide_compuestasCateg_ggplot2}
\title{DemBas_piramide_compuestasCateg_ggplot2}
\usage{
DemBas_piramide_compuestasCateg_ggplot2(
  datosPiramide,
  porcentajes = TRUE,
  etiquetas = FALSE,
  etiquetas.size = 4,
  colores = NULL,
  ordeninverso = FALSE,
  alfa = 1,
  bar.size = 1,
  etiq.hombre = "Hombre",
  etiq.mujer = "Mujer"
)
}
\arguments{
\item{etiq.mujer}{}
}
\description{
DemBas_piramide_compuestasCateg_ggplot2
}
