% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piramidespoblacion.R
\name{DemBas_piramide_ggplot2_linea}
\alias{DemBas_piramide_ggplot2_linea}
\title{DemBas_piramide_ggplot2_linea}
\usage{
DemBas_piramide_ggplot2_linea(
  datosPiramide,
  porcentajes = TRUE,
  etiquetas = FALSE,
  etiquetas.size = 4,
  UsaCaso = FALSE,
  etiq.hombre = "Hombre",
  etiq.mujer = "Mujer",
  colorear = "Sexo",
  colores = NULL
)
}
\arguments{
\item{datosPiramide}{}

\item{porcentajes}{}

\item{etiquetas}{}

\item{etiquetas.size}{}

\item{UsaCaso}{}

\item{etiq.hombre}{}

\item{etiq.mujer}{}

\item{colorear}{}

\item{colores}{}
}
\description{
DemBas_piramide_ggplot2_linea
}
\examples{
# Ejecutar ejemplo de función DemBas_piramide_ggplot2()
DemBas_piramide_ggplot2_linea(dfPir2017,colorear = "Sexo",
                            etiq.hombre = "Hombres",etiq.mujer = "Mujeres") + 
  labs(title = "Perfil de la Pirámide de Población de España en 2017") +
  scale_x_discrete(
    # si la variable edad fuera numeric debería usarse scale_x_continuous
    breaks = seq(0,105,by=5),
    labels = paste0(as.character(seq(0,105,by=5)), "")) +
  guides(colour="none")
}
