% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piramidespoblacion.R
\name{DemBas_piramide_superpuestas_ggplot2}
\alias{DemBas_piramide_superpuestas_ggplot2}
\title{DemBas_piramide_superpuestas_ggplot2}
\usage{
DemBas_piramide_superpuestas_ggplot2(
  datosPiramide,
  porcentajes = TRUE,
  etiquetas = FALSE,
  etiquetas.size = 4,
  colores = NULL,
  transparente = FALSE,
  alfa = 0.1,
  bar.size = 1,
  etiq.hombre = "Hombre",
  etiq.mujer = "Mujer",
  porc.X.ini = 0,
  porc.X.by = 0.2
)
}
\arguments{
\item{etiq.mujer}{}
}
\description{
DemBas_piramide_superpuestas_ggplot2
}
\examples{
# Ejecutar el código ejemplo de la función DemBas_piramide_ggplot2
dfPir2002 = dfej02a \%>\%
  dplyr::filter( Periodo=="1 de enero de  2002",
                 !(Sexo=="Ambos sexos"),
                 !(Edad.simple \%in\% c("100 y más años","Total"))) \%>\%
  dplyr::select(Edadchar=Edad.simple,
                Sexo,
                Poblacion = value)
dfPir2002$Edad = DemBas_extrae_codigo_provincia(dfPir2002$Edadchar)
dfPir2002$Edad = factor(dfPir2002$Edad,levels =unique(dfPir2002$Edad))
dfPir2002$Poblacion[is.na(dfPir2002$Poblacion)] = 0
dfPir2002y2017 = rbind(dfPir2002,dfPir2017)
dfPir2002y2017$Caso = c(rep(2002,nrow(dfPir2002)),rep(2017,nrow(dfPir2017)))
head(dfPir2002y2017)
DemBas_piramide_superpuestas_ggplot2(dfPir2002y2017,
                                   etiq.hombre = "Hombres",etiq.mujer = "Mujeres",
                                   transparente = T) +
  labs(title = "Pirámides de Población de España en 2002 y 2017 superpuestas") +
  scale_x_discrete(
    breaks = seq(0,105,by=5),
    labels = paste0(as.character(seq(0,105,by=5)), ""))
}
