% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piramidespoblacion.R
\name{DemBas_piramidePorc}
\alias{DemBas_piramidePorc}
\title{DemBas_piramidePorc}
\usage{
DemBas_piramidePorc(
  datosPiramide,
  Gtitulo = "Pirámide Población de la provincia de Sevilla",
  Gsubtitulo = "Año 2020",
  Gtitulo.X = "Porcentajes",
  GHombresEtiq = "Hombres",
  GMujeresEtiq = "Mujeres",
  Gedadfinal = 100,
  Gext_izq = -4.5,
  Gext_der = 4.5,
  Glimite = 0.5,
  Gsizeletra = 2.5,
  GSegmentos = TRUE
)
}
\arguments{
\item{datosPiramide}{}

\item{Gtitulo}{}

\item{Gsubtitulo}{}

\item{Gtitulo.X}{}

\item{GHombresEtiq}{}

\item{GMujeresEtiq}{}

\item{Gedadfinal}{}

\item{Gext_izq}{}

\item{Gext_der}{}

\item{Glimite}{}

\item{Gsizeletra}{}

\item{GSegmentos}{}
}
\value{
ggplot2 graphics de la pirámide de Población con porcentajes
}
\description{
DemBas_piramidePorc
}
\examples{
load(file = system.file("examples/04003px.RData", package = "DemographyBasic"))

ano_selec = 2020
Espanoles_Extranjeros = "Españoles" 
CCAA_Prov = "Sevilla"

datosPiramide =  datos |> 
  dplyr::filter(Ano == ano_selec & 
                  Sexo \%in\% c("Mujeres", "Hombres") & 
                  Edad != "TOTAL" & 
                  CCAA.Prov == CCAA_Prov &
                  Espanoles.Extranjeros == Espanoles_Extranjeros) |> 
  dplyr::rename(Poblacion = value) |> 
  dplyr::select(Edad, Sexo, Poblacion) 
(g_pir1 = DemBas_piramidePorc(datosPiramide, 
                            Gtitulo = "Pirámide Población de la provincia de Sevilla",
                            Gsubtitulo = "Año 2020  (españoles)",
                            GSegmentos = FALSE))
                            
}
