% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piramidespoblacion.R
\name{DemBas_piramidePorc_Generaciones}
\alias{DemBas_piramidePorc_Generaciones}
\title{DemBas_piramidePorc_Generaciones}
\usage{
DemBas_piramidePorc_Generaciones(
  pop3,
  Ano_ref = 2020,
  Gtitulo = "Pirámide Población de la provincia de Sevilla",
  Gsubtitulo = "Año 2020",
  Gtitulo.X = "Porcentajes",
  GHombresEtiq = "Hombres",
  GMujeresEtiq = "Mujeres",
  Gedadfinal = 100,
  Gext_izq = -4.5,
  Gext_der = 4.5,
  Glimite = 0.5,
  Gsizeletra = 2.5,
  GpresentaResumen = TRUE,
  GSegmentos = TRUE,
  GHombresColor = "#5BB4E5",
  GMujeresColor = "#DE61D8",
  GSegmentosColor = "#00ff00"
)
}
\arguments{
\item{pop3}{data.frame o tibble with columns: Edad, Sexo, Porcentajes}

\item{GSegmentosColor}{}
}
\description{
DemBas_piramidePorc_Generaciones
}
\examples{
load(file = system.file("examples/pop3.RData", package = "DemographyBasic"))
g_pir3gen = DemBas_piramidePorc_Generaciones(pop3)
g_pir3gen
ggsave("piramide.png", g_pir3gen, width = 12, height = 10)
png("piramide2.png", width = 1200)
print(g_pir3gen)
dev.off()
}
