% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piramidespoblacion.R
\name{DemBas_piramides_enfrentadas_ggplot2}
\alias{DemBas_piramides_enfrentadas_ggplot2}
\title{DemBas_piramides_enfrentadas_ggplot2}
\usage{
DemBas_piramides_enfrentadas_ggplot2(
  datosPiramide,
  porcentajes = TRUE,
  etiquetas = FALSE,
  etiquetas.size = 4,
  UsaCaso = TRUE,
  etiq.hombre = "Hombre",
  etiq.mujer = "Mujer",
  colorear = "Sexo",
  colores = NULL,
  nfilas = NULL,
  ncols = NULL
)
}
\arguments{
\item{ncols}{}
}
\description{
DemBas_piramides_enfrentadas_ggplot2
}
