% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presentardf.R
\name{DemBas_presentadf}
\alias{DemBas_presentadf}
\title{DemBas_presentadf}
\usage{
DemBas_presentadf(
  datos1,
  scaption = NULL,
  apaisadalatex = FALSE,
  variaspaginas = TRUE,
  fuentesize = NULL,
  CompletaAncho = FALSE
)
}
\arguments{
\item{datos1}{}

\item{scaption}{}

\item{apaisadalatex}{}

\item{variaspaginas}{}

\item{fuentesize}{}

\item{CompletaAncho}{}
}
\value{
devuelve una tabla para ser presentada en formato LaTeX o HTML
}
\description{
Presenta un data.frame en formato LaTeX o HTML con kableExtra
}
\examples{
DemBas_presentadf(mtcars)
}
