% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presentardf.R
\name{DemBas_presentadf_modelo01}
\alias{DemBas_presentadf_modelo01}
\title{DemBas_presentadf_modelo01}
\usage{
DemBas_presentadf_modelo01(
  df_tabla,
  tablavariaspaginas = T,
  tfuente = 8,
  leyendatabla = NULL,
  vCabecera = NULL,
  valigncol = NULL
)
}
\arguments{
\item{valigncol}{}
}
\description{
Presenta un data.frame en formato LaTeX o HTML con kableExtra
}
\examples{
DemBas_presentadf_modelo01(mtcars)
}
