% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxR2.R
\name{DemBas_read_px}
\alias{DemBas_read_px}
\title{DemBas_read_px}
\usage{
DemBas_read_px(
  filename,
  encoding = NULL,
  na.strings = c("\\".\\"", "\\"..\\"", "\\"...\\"", "\\"....\\"", "\\":\\"")
)
}
\arguments{
\item{na.strings}{}
}
\value{
data.frame
}
\description{
read a file px (INE, Eurostat, etc) and return a data.frame
}
\examples{
# example code
# Importing a .px file
example_data_px <- DemBas_read_px(system.file("examples/56940.px", package = "DemographyBasic"))
head(example_data_px)
}
