% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxR2.R
\name{DemBas_read_px_encodingIECA}
\alias{DemBas_read_px_encodingIECA}
\title{DemBas_read_px_encodingIECA}
\usage{
DemBas_read_px_encodingIECA(ficheropx)
}
\arguments{
\item{ficheropx}{character with the name of the file px}
}
\value{
data.frame with the data
}
\description{
read a file px (IECA) and return a data.frame
}
\examples{
d103965windows4 = DemBas_read_px_encodingIECA("ieca_export_103965.px")
}
