% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presentardf.R
\name{DemBas_redondear}
\alias{DemBas_redondear}
\title{DemBas_redondear}
\usage{
DemBas_redondear(x, digitos = 0)
}
\arguments{
\item{x}{vector numérico}

\item{digitos}{número de decimales}
}
\value{
vector numérico redondeado
}
\description{
Redondea un vector numérico
}
\examples{
v1 = DemBas_redondear(rnorm(20), 2)
v1
}
