% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presentardf.R
\name{DemBas_redondear_df}
\alias{DemBas_redondear_df}
\title{DemBas_redondear_df}
\usage{
DemBas_redondear_df(df, digitos = 0)
}
\arguments{
\item{df}{data.frame a redondear}

\item{digitos}{número de decimales}
}
\value{
data.frame con las columnas numéricas redondeadas
}
\description{
Redondea las columnas numéricas de un data.frame
}
\examples{
df = data.frame(nombres = sample(letters,20), x = rnorm(20), y = rnorm(20))
df2 = DemBas_redondear_df(df, 2)
}
