% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxR2.R
\name{DemBas_url_px_IECA}
\alias{DemBas_url_px_IECA}
\title{DemBas_url_px_IECA}
\usage{
DemBas_url_px_IECA(id, ficheropx)
}
\arguments{
\item{id}{character with the id of the px file in the IECA web}

\item{ficheropx}{character with the name of the file px to save}
}
\value{
character with the url of the file px
}
\description{
Descarga un fichero px de la web del IECA a partir de su identificador
}
\examples{
url1 = DemBas_url_px_IECA("103965", "ieca_export_103965_ff.px")
df103965_ff = DemBas_read_px_encodingIECA("ieca_export_103965_ff.px")
}
