% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxR2.R
\name{DemBas_url_px_INE}
\alias{DemBas_url_px_INE}
\title{DemBas_url_px_INE}
\usage{
DemBas_url_px_INE(id, ficheropx)
}
\arguments{
\item{id}{character with the id of the px file in the INE web}

\item{ficheropx}{character with the name of the file px to save}
}
\value{
character with the url of the file px
}
\description{
Descarga un fichero px de la web del INE a partir de su identificador
}
\examples{
url1 = DemBas_url_px_INE("2855", "2855_ine.px")
}
